_count_class_lookup should be a classmethod
Since it's a cached method, every invocation of this method is preceded
by a cache lookup. If it's a bound method, then the cache lookup must
take into consideration the `self` argument as part of the invocation.
It must then call the self.__eq__(other) method to verify if it exists
in the cache. However, since _count_class_lookup is used in the
constructor to define self._cov, and since self._cov is part of the
equality check, then by the time self.__eq__(other) is called, self._cov
would not have been initialized yet, and the invocation would throw an
AttributeError (since self._cov is not yet defined).