Fixed bug in TCPChannel
When debugging PtraceChannel, it is often the case that ignore_callbacks
are replaced by `lambda x: False` to inspect the sequence of syscalls
executed by the target. However, that means that syscall callbacks then
receive all syscalls regardless of the underlying assumption about the
ignore function. In this case, the underlying assumption was broken and
the syscall callback mishandled it. In the case that the syscall name
does not match any of the listened-to names, the callback should do
nothing; but in this case, it continued execution past the if statement
and injected the forkserver at an unexpected wrong location.