current_path is now maintained by StateManagerContext
It is still ambiguous whether this is a StateManager concern or a loader concern. From the SMan's pov, it does not matter how the state was reached as long as it can be accurately and reliably reproduced; for a VM-snapshot-based loader, it may suffice to just specify the VM snapshot image and the last_input. However, practically, we'd still like a cold-start PoC, regardless of the nature of the loader, making it the SMan's responsibility to track what has been sent to the target in order to reproduce it later.