Removed async_enumerate where not necessary
async_enumerate is an async generator, which, when iterated on, results
in scheduling finalizers to the event loop. On the critical path, these
tasks take up a huge chunk of the time in I/O processing, and the
alternative here is just a loop counter.

Note: StateManagerContext now iterates over inputs as normal generators;
if, in the future, inputs require async generators (e.g. generating an
input in real time requires I/O waiting), the for loop must be replaced
again by async for.