Force IPv4 in WebUI
By default, docker does not support IPv6 unless explicitly specified in
the daemon config. When loop.create_server is called, it resolved the
(host, port) tuple with socket.getaddrinfo, which returns a list of
possible socket configurations, including an IPv6 tuple.

Currently, aiohttp.web.TCPSite does not support specifying the address
family of the socket, so we get around that by specifying '127.0.0.1' to
explicitly say we want IPv4 localhost. This may break on systems where
the loopback address is not 127.0.0.1.