Added option to chunk-up generated instructions
Since an Instruction is the "atomic" fuzzing operation, the fuzzer may
lose a lot of progress by generating N-byte instructions whose first
few M bytes are valid and the remaining N-M bytes are rejected. Normally
the fuzzer would discard the whole instruction, since it cannot dissect
it without making assumptions about the nature of the instruction.

The HavocMutator is thus the first component that controls the nature of
instructions it generates (arguably, the InputGenerators; but for now,
they are all based on HavocMutator). So, to work around this, we can
split up generated data into multiple chunks and store each in a separate
instruction. As such, the most progress we could lose is chunk_size,
instead of M bytes.