Removed incorrect assertion
If the tracker dissolves a state, the old capability matrix would still
carry information about that state. However, `from_idx` would only
refer to old states that still exist in the current graph. Hence, the
dimensions of the cap matrix are only guaranteed to be >= the size of
from_idx.