Allow CoverageTracker.current_state to return seen states
This change allows the tracker to check if the observed coverage can be
attributed to a seen snapshot, instead of returning the default source.
This allows the fuzzer to discover multiple inputs for the same snapshot
transition.

To improve this detection, a DFS search can be done to attribute
feedback to multiple transitions instead of just one.